% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mbleven_compare}
\alias{mbleven_compare}
\title{Perform mbleven comparison on two strings}
\usage{
mbleven_compare(str1, str2)
}
\arguments{
\item{str1}{A character(1) vector.}

\item{str2}{A character(1) vector.}
}
\description{
This is an Rcpp implementation of the mbleven algorithm. It takes two
strings as input and returns 0, 1, or 2 when the edit distance of the
strings are 0, 1, or 2 repsecitvely. If the edit distance if over three
edits, the algoirthm will return -1.
}
\examples{
mbleven_compare('aaa', 'aaa')
mbleven_compare('aaa', 'abb')
mbleven_compare('aaa', 'bbb')
}
